Moved test class to new file called 'test_data_zarr'
